refactor(dispatch): enforce canonical five module ownership
