+++ rust/crates/sky_player_rs/src/engine/worker/dispatch/mod.rs
@@ -17,21 +17,8 @@
 //!   publication.
 
 mod authored;
-#[cfg(not(any(test, feature = "test-support")))]
-mod observer;
-#[cfg(any(test, feature = "test-support"))]
 pub(crate) mod observer;
-#[cfg(not(any(test, feature = "test-support")))]
-mod observer_drain;
-// `pub(crate)` under test/test-support so engine.rs's public §8.11/§8.12
-// bridge can reach the queue primitives and hooks through a crate-visible
-// path (still no crate-wide module leak in production).
-#[cfg(any(test, feature = "test-support"))]
-pub(crate) mod observer_drain;
 mod release;
-#[cfg(not(any(test, feature = "test-support")))]
-mod timing;
-#[cfg(any(test, feature = "test-support"))]
 pub(crate) mod timing;
 
 /// Outcome of one dispatch-loop authored or pending-release step.
@@ -97,9 +84,7 @@ pub(crate) struct AuthoredBatchView {
 pub(super) type BatchViewResult = Result<Option<AuthoredBatchView>, DispatchStep>;
 
 pub(crate) use authored::dispatch_authored_packet;
-pub(crate) use observer_drain::{
-    PendingObservationQueue, drain_one_observer, observer_has_safe_slack,
-};
+pub(crate) use observer::{PendingObservationQueue, drain_one_observer, observer_has_safe_slack};
 pub(crate) use release::{PendingReleaseContext, dispatch_due_pending_releases};
 
 // `observer_initial_budget_override_us` is reached crate-internally via
@@ -107,7 +92,7 @@ pub(crate) use release::{PendingReleaseContext, dispatch_due_pending_releases};
 // queue primitives + §8.12 hooks are re-exported at crate root under
 // `engine::dispatch_primitives` / `engine::observer_test_hooks` instead.
 #[cfg(any(test, feature = "test-support"))]
-pub(super) use observer_drain::observer_initial_budget_override_us;
+pub(super) use observer::observer_initial_budget_override_us;
 
 use super::super::{ActionKind, LatencyClass, QpcTicks, TimelineTicks};
 use super::DispatchPath;
