+++ rust/crates/sky_player_rs/src/engine/worker/planning.rs
@@ -5,28 +5,19 @@
 //! release lead, and the earliest wait deadline so prepare-due and wait-until
 //! cannot disagree on lead selection.
 
-use super::health::{DispatchLeadEstimate, DispatchPath, estimate_dispatch_path_lead};
+pub(crate) use super::dispatch::timing::{
+    AuthoredDispatchPlan, next_authored_path, pending_lead_for_polyphony, resolve_authored_lead,
+    startup_lead_for_first_packet,
+};
+use super::health::DispatchPath;
 use crate::engine::config::TimingOptions;
 use sky_dispatch_core::coordinator::{
-    CoordinatorError, PendingDispatchPlan, RuntimeDispatchCoordinator, physical_packet_kind,
+    CoordinatorError, PendingDispatchPlan, RuntimeDispatchCoordinator,
 };
-use sky_dispatch_core::estimator::{LatencyClass, SendLatencyEstimator, SendPath};
-use sky_dispatch_core::model::PhysicalPacketKind;
+use sky_dispatch_core::estimator::{LatencyClass, SendLatencyEstimator};
 use sky_dispatch_core::time::{DurationTicks, TimelineTicks};
 use sky_dispatch_win32::clock::QpcClock;
 use std::fmt;
-
-/// Snapshot of the next authored packet's dispatch path and lead.
-///
-/// Built once per worker loop epoch and reused for both
-/// `prepare_next_due_authored` and `next_deadline_ticks`.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(crate) struct AuthoredDispatchPlan {
-    pub(crate) path: DispatchPath,
-    pub(crate) lead_us: u64,
-    pub(crate) lead_ticks: DurationTicks,
-    pub(crate) lead_saturated: bool,
-}
 
 /// Immutable plan for one worker loop epoch.
 ///
@@ -107,86 +98,6 @@ impl From<CoordinatorError> for PlanningError {
     }
 }
 
-/// Resolve the path-aware lead for one authored packet without reading QPC.
-fn resolve_authored_lead(
-    estimator: &SendLatencyEstimator,
-    path: DispatchPath,
-    latency_class: LatencyClass,
-    timing: &TimingOptions,
-    enable_adaptive_lead: bool,
-) -> DispatchLeadEstimate {
-    if timing.dispatch_lead_us > 0 {
-        return DispatchLeadEstimate {
-            applied_us: timing.dispatch_lead_us,
-            saturated: false,
-        };
-    }
-    if !enable_adaptive_lead {
-        return DispatchLeadEstimate {
-            applied_us: 0,
-            saturated: false,
-        };
-    }
-    estimate_dispatch_path_lead(
-        estimator,
-        path,
-        latency_class,
-        timing.strict_timing,
-        timing.max_lead_us,
-    )
-}
-
-/// Classify the next authored packet into a [`DispatchPath`].
-///
-/// Empty physical masks (stale Up suppression metadata) keep the historical
-/// Down-polyphony fallback so wait/prepare stay consistent with prior behavior.
-fn next_authored_path(coordinator: &RuntimeDispatchCoordinator) -> Option<DispatchPath> {
-    let (up_mask, down_mask) = coordinator.next_authored_packet_masks()?;
-    let up_count = up_mask.count_ones() as usize;
-    let down_count = down_mask.count_ones() as usize;
-    match physical_packet_kind(up_mask, down_mask) {
-        Ok(PhysicalPacketKind::UpOnly) => Some(DispatchPath::UpOnly { up_count }),
-        Ok(PhysicalPacketKind::DownOnly) => Some(DispatchPath::DownOnly { down_count }),
-        Ok(PhysicalPacketKind::Mixed) => Some(DispatchPath::Mixed {
-            up_count,
-            down_count,
-        }),
-        Err(_) => {
-            let polyphony = coordinator.next_authored_polyphony().max(1);
-            Some(DispatchPath::DownOnly {
-                down_count: polyphony,
-            })
-        }
-    }
-}
-
-fn pending_lead_for_polyphony(
-    estimator: &SendLatencyEstimator,
-    qpc_clock: QpcClock,
-    polyphony: usize,
-    latency_class: LatencyClass,
-    timing: &TimingOptions,
-    enable_adaptive_lead: bool,
-) -> Result<(DurationTicks, bool), PlanningError> {
-    let (lead_us, saturated) = if timing.dispatch_lead_us > 0 {
-        (timing.dispatch_lead_us, false)
-    } else if enable_adaptive_lead {
-        let estimate = estimator.estimate_lead_with_class_and_policy(
-            SendPath::UpOnly,
-            polyphony,
-            latency_class,
-            timing.strict_timing,
-        );
-        (estimate.applied_us, estimate.saturated)
-    } else {
-        (0, false)
-    };
-    qpc_clock
-        .duration_from_us(lead_us)
-        .map(|ticks| (ticks, saturated))
-        .map_err(|error| PlanningError::TimeConversion(format!("{error:?}")))
-}
-
 /// Build the single immutable dispatch plan for the current worker loop epoch.
 ///
 /// Contract:
@@ -231,10 +142,7 @@ pub(crate) fn plan_next_dispatch(
             timing,
             enable_adaptive_lead,
         )
-        .map_err(|error| match error {
-            PlanningError::TimeConversion(message) => CoordinatorError::TimeConversion(message),
-            PlanningError::Coordinator(inner) => inner,
-        })
+        .map_err(|error| CoordinatorError::TimeConversion(format!("{error:?}")))
     })?;
 
     let authored_lead_ticks = authored
@@ -248,26 +156,6 @@ pub(crate) fn plan_next_dispatch(
         pending,
         deadline_ticks,
     })
-}
-
-/// Resolve the path-aware lead used to anchor the startup wait before the
-/// first main-loop `NextDispatchPlan` is built.
-///
-/// The normal loop derives lead from the next authored packet's path; startup
-/// must use the same policy instead of a hard-coded Down lead, otherwise a
-/// first authored `UpOnly`/`Mixed` packet anchors its physical boundary with
-/// the wrong directional lead.
-pub(crate) fn startup_lead_for_first_packet(
-    coordinator: &RuntimeDispatchCoordinator,
-    estimator: &SendLatencyEstimator,
-    latency_class: LatencyClass,
-    timing: &TimingOptions,
-    enable_adaptive_lead: bool,
-) -> u64 {
-    let path = next_authored_path(coordinator).unwrap_or_else(|| DispatchPath::DownOnly {
-        down_count: coordinator.next_authored_polyphony().max(1),
-    });
-    resolve_authored_lead(estimator, path, latency_class, timing, enable_adaptive_lead).applied_us
 }
 
 #[cfg(test)]
